optimize mathquill updates for sliders and simple numbers
We check if the request latex and the previous latex are in a form that we can efficiently and
safely mutate. Basically, we look for a number at the end of the latex. If both the previous and the
current latex have the same prefix before that number then we can skip to just updating the number.

The minus sign adds a ton of code complexity here. I decided to implement it 'correctly' because the
PlusMinus symbol is different than the Digit symbol. It has logic to decide if it's a binary
operator, and it could possibly add even more functionality. It probably isn't a good idea to
repurpose a PlusMinus node for a Digit node.

Once we deal with adding or removing the minus symbol, we start repurposing as many nodes as
possible by simply writing new digits in them. After that we decide if we need to remove any extra
symbols. If so, we do. And then we check if we need to add any symbols. If so, we do.

We take care to keep the tree in a completely valid state after the update. The cost savings come
from a few parts:
1) we don't have to parse more than just the ending number
2) we don't have to create and append more nodes than necessary
3) the nodes we create are created efficiently. Mathquill uses the jquery dom parser and its slow.

At the very end we check that when we read the mathquill out we get the latex we intended. This is
a quick sanity check. If the results do not check out then we resort to a full update.

This is pretty much just as fast as our ugly transient-mathquill hacks but has the added benefit
we do not need to put that special logic anywhere. It's baked into mathquill now. So draggable table
columns and sliders are both fast for free. The second added benefit is that selection works
correctly for these mathquills. You can mouse down in an animated slider and start selecting exactly
where you moused down. Before, the best we could do is put focus at the end of the expression.