harcode dependency order in Makefile
There are differences in the order that files get included when you compare the built file on linux and mac. I think the order that glob returns files differs between operating systems. We are now explicitly listing the files to avoid random differences. We only stumbled upon this because there is actually a bug that exists when mathquill is compiled on linux but doesn't exist when compiled on mac.  We are accepting the order that mac uses. There's no guarantee that this is the definite best order. Only time will tell.