Use regular expression to separate digits for mathspeak
The previous version added a space before the decimal point so that most screen readers would announce digits after the decimal individually (mostly a problem on the Mac). This worked for nearly every device except iOS devices.

This addition replaces the previous hack with a regexp replace that splits out all contiguous digits after the decimal which should work everywhere.