Move mixed fraction speech determination to fraction mathspeak method
Instead of just checking whether the previous mathspeak character was a number, we now scan leftward from the fraction element itself and insert the and conjunction if all items to the left of the current block are digits, the plus or minus sign, or whitespace.

In particular, this stops Mathquill from reporting 'and' for decimals followed by fractions.